pyproject.toml
[tool.poetry]
name = "identitytheft-bot"
version = "0.1.0"
description = ""
authors = ["Sakari Marttinen"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
Telethon = "^1.38.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
